backend/requirements.txt
openai
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv